Documentation: Minor: Clarify creating custom mappings
> I believe existing version of README.md is not clear enough. For example it says that you need to do custom mappings before sourcing the plugin's script. But in fact you need to assign 1 to `g:mw_no_warnings` before sourcing the script. In this case it does not matter at which point to do those custom mappings (although I prefer to do that after sourcing the script).
> In this PR I'm trying to clarify this topic in README.md

Right; it wasn't explicitly stated that g:mw_no_mappings needs to be set before the plugin is loaded (whereas most other config variable changes also affect the running plugin).
The flag is a convenience functionality as the plugin has so many default mappings that it's cumbersome to custom-map all of them if you just want to use a few. Though it's correct that with the flag set, it doesn't matter any longer when the used mappings are overridden, in the general case (i.e. without the flag set), this still is important.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: Vim 9.1.1029
Co-authored-by: Ingo Karkat <swdev@ingo-karkat.de>